Move some file into component
