+++ frontend/app/business/profile/page.tsx
@@ -7,14 +7,14 @@ import { useEffect, useState } from 'react';
 import Image from 'next/image';
 import { toast } from "react-toastify";
 import BusinessNavbar from '../components/BusinessNavbar';
-import WeeklyEntryChart from '../WeeklyEntryChart';
-import EntryTimeChart from '../EntryTimeChart';
-import EstimateTimeChart from '../EstimateTimeChart';
-import EstimateDayChart from '../EstimateDayChart';
-import EstimateQueueChart from '../EstimateQueueChart';
-import EntryQueueChart from '../EntryQueueChart';
-import EntryChart from '../EntryChart';
-import TopQueue from '../TopQueue';
+import WeeklyEntryChart from '../components/chart/WeeklyEntryChart';
+import EntryTimeChart from '../components/chart/EntryTimeChart';
+import EstimateTimeChart from '../components/chart/EstimateTimeChart';
+import EstimateDayChart from '../components/chart/EstimateDayChart';
+import EstimateQueueChart from '../components/chart/EstimateQueueChart';
+import EntryQueueChart from '../components/chart/EntryQueueChart';
+import EntryChart from '../components/chart/EntryChart';
+import TopQueue from '../components/TopQueue';
 import ApiProxy from '@/app/api/proxy';
 import { useRouter } from 'next/navigation';
 import { useAuth } from '@/components/AuthProvider';
